+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - [refactor:refactor Polaris registration.](https://github.com/Tencent/spring-cloud-tencent/pull/1055)
 - [fix:fix SCG report wrong service bug when using IP routing.](https://github.com/Tencent/spring-cloud-tencent/pull/1063)
 - [fix:fix gray release examples bug.](https://github.com/Tencent/spring-cloud-tencent/pull/1066)
+- [fix:fix router label feign interceptor order.](https://github.com/Tencent/spring-cloud-tencent/pull/1069)